requirements.txt
PyQt6
psycopg2-binary
python-dotenv
bcrypt
